refactor examples tests
